created HandlePagePassword componenet
