+++ .vscode/settings.json
@@ -1,5 +1,7 @@
 {
   "cSpell.words": [
-    "dlcm"
+    "bandcamp",
+    "dlcm",
+    "soundcloud"
   ]
 }
